Implement dynamic VRPTW simulation and refactor distance calculations
- Added a new test script for dynamic settings in `notebooks/dynamic_setting/test.py`.
- Refactored distance calculations in `normalization.py` to use integers for scale.
- Updated `solomon_parser.py` to apply distance scaling to depot and customer attributes.
- Modified `arrivals.py` to remove debug print statements.
- Enhanced `dynamic_instance.py` to include metadata for dynamic nodes and improve customer handling.
- Refactored `simulator.py` to utilize integer distances and streamline vehicle travel calculations.
- Updated `snapshot.py` to change float attributes to integers for vehicle state.
- Enhanced `ortools_dynamic.py` to support soft time windows and improve distance callbacks.
- Adjusted `objective.py` to calculate total distance using dynamic instance routes.
- Added unit tests in `test_dynamic_instance.py` to validate dynamic instance behavior and distance calculations.

+++ src/dvrptw_bench/data/normalization.py
@@ -7,9 +7,9 @@ import math
 from dvrptw_bench.common.typing import Node
 
 
-def euclidean(a: Node, b: Node, scale: float = 1.0) -> float:
-    return scale * math.hypot(a.x - b.x, a.y - b.y)
+def euclidean(a: Node, b: Node, scale: int = 1) -> float:
+    return scale * int(math.hypot(a.x - b.x, a.y - b.y))
 
 
-def distance_matrix(nodes: list[Node], scale: float = 1.0) -> list[list[float]]:
+def distance_matrix(nodes: list[Node], scale: int = 1) -> list[list[float]]:
     return [[euclidean(i, j, scale=scale) for j in nodes] for i in nodes]
